+++ spec/client/paginated_enumeration_spec.rb
@@ -1,6 +1,220 @@
 require 'securerandom'
 require 'spec_helper.rb'
 
+#
+# These tests define the following Services.
+#
+# Clients can call into any of them to invoke the different calling semantics
+# between them.
+#
+#
+# ┌──────────────────────────────────────────────┐ ┌──────────────────────────┐
+# │                                              │ │                          │
+# │               RSpecNumberService             │ │ RSpecRemoteNumberService │
+# │                                              │ │                          │
+# │                                              │ │                          │
+# │ ┌──────────────┐           ┌────────────────┐│ │  ┌───────────────────┐   │
+# │ │              │  inter    │                ││ │  │                   │   │
+# │ │ RSpecNumber  │◀resource ─│RSpecEvenNumber ││ │  │  RSpecOddNumber   │   │
+# │ │              │  local    │                ││ │  │                   │   │
+# │ └──────────────┘           └────────────────┘│ │  └───────────────────┘   │
+# │         ▲                                    │ │            │             │
+# │         │                            inter   │ │            │             │
+# │         └───────────────────────────resource ┼─┼────────────┘             │
+# │                                      remote  │ │                          │
+# └──────────────────────────────────────────────┘ └──────────────────────────┘
+#
+# To start the services in your specs do:
+#
+#     spec_helper_start_svc_app_in_thread_for( RSpecNumberService )
+#     spec_helper_start_svc_app_in_thread_for( RSpecRemoteNumberService)
+#
+
+################################################################################
+#
+# Create a 'RSpecNumber' Resource with the following properties:
+#
+# - manages 'Number' resources ie: { 'number': 3 }, for numbers between 0 & 999
+# - provides a public 'list' endpoint (no session needed)
+# - pagination
+# - will generate an error when asked to retrieve any 'Number' resource with a
+#   number value >= 500 && filter_data['force_error'] is set (to anything)
+#
+class RSpecNumberImplementation < Hoodoo::Services::Implementation
+
+  public
+
+  # Number resources are all in this range
+  NUMBER_RANGE = 0..999
+
+  # Number resources that generate errors are all in this range
+  ERROR_RANGE  = 500..999
+
+  def list( context )
+    request  = context.request
+
+    resources = []
+    implode = false
+    0.upto( request.list.limit - 1 ) do |i|
+      num = request.list.offset + i
+      implode = implode || ERROR_RANGE.include?( num )
+      if NUMBER_RANGE.include?( num )
+        resources << { 'number' => num }
+      else
+        break
+      end
+    end
+
+    context.response.set_resources( resources, resources.count )
+    if implode && request.list.filter_data[ 'force_error' ]
+      context.response.add_error( 'platform.malformed' )
+    end
+  end
+
+end
+
+#
+# Interface for our implementation
+#
+class RSpecNumberInterface < Hoodoo::Services::Interface
+  interface :RSpecNumber do
+    endpoint       :numbers, RSpecNumberImplementation
+    to_list do
+      filter :force_error
+    end
+    public_actions :list
+  end
+end
+
+
+################################################################################
+#
+# Create a 'RSpecEvenNumber' Resource with the following properties:
+#
+# - Calls RSpecNumber via the 'inter_resource_local' calling mechanism
+# - Only returns 'even' numbers, 0, 2, 4... for numbers between 0 & 999
+# - provides a public 'list' endpoint (no session needed)
+#
+# See RSpecNumberImplementation for error handling etc
+#
+class RSpecEvenNumberImplementation < Hoodoo::Services::Implementation
+
+  public
+
+  def list( context )
+    request   = context.request
+    endpoint  = context.resource( :RSpecNumber, 1 )
+    resources = []
+    limit     = request.list.limit  ? request.list.limit  : 50
+    offset    = request.list.offset ? request.list.offset : 0
+
+    # We always iterate through every Number resource - yeah its dumb
+    endpoint.list( { :filter => request.list.filter_data } ).enumerate_all do | number_res |
+
+      if number_res.platform_errors.has_errors?
+        context.response.add_errors( number_res.platform_errors )
+        break
+      end
+
+      number = number_res['number']
+
+      # Number in the correct range & is 'even'
+      resources << number_res if number >= ( offset * 2 ) && number.even?
+      break if resources.size >= limit
+
+    end
+
+    context.response.set_resources( resources, resources.count )
+  end
+
+end
+
+#
+# Interface for our implementation
+#
+class RSpecEvenNumberInterface < Hoodoo::Services::Interface
+  interface :RSpecEvenNumber do
+    endpoint       :even_numbers, RSpecEvenNumberImplementation
+    to_list do
+      filter :force_error
+    end
+    public_actions :list
+  end
+end
+
+################################################################################
+#
+# Define our service, that implements both resources
+#
+class RSpecNumberService < Hoodoo::Services::Service
+  comprised_of RSpecNumberInterface,
+               RSpecEvenNumberInterface
+end
+
+
+################################################################################
+#
+# Create a 'RSpecOddNumber' Resource with the following properties:
+#
+# - Calls RSpecNumber via the 'inter_resource_remote' calling mechanism
+# - Only returns 'odd' numbers, 1, 3, 5 ... for numbers between 0 & 999
+# - provides a public 'list' endpoint (no session needed)
+#
+# See RSpecNumberImplementation for error handling etc
+#
+class RSpecOddNumberImplementation < Hoodoo::Services::Implementation
+
+  public
+
+  def list( context )
+    request   = context.request
+    endpoint  = context.resource( :RSpecNumber, 1 )
+    resources = []
+    limit     = request.list.limit  ? request.list.limit  : 50
+    offset    = request.list.offset ? request.list.offset : 0
+
+    # We always iterate through every Number resource - yeah its dumb
+    endpoint.list( { :filter => request.list.filter_data } ).enumerate_all do | number_res |
+
+      if number_res.platform_errors.has_errors?
+        context.response.add_errors( number_res.platform_errors )
+        break
+      end
+
+      number = number_res['number']
+
+      # Number in the correct range & is 'odd'
+      resources << number_res if number >= ( offset * 2 ) && number.odd?
+      break if resources.size >= limit
+
+    end
+
+    context.response.set_resources( resources, resources.count )
+  end
+
+end
+
+#
+# Interface for our implementation
+#
+class RSpecOddNumberInterface < Hoodoo::Services::Interface
+  interface :RSpecOddNumber do
+    endpoint       :odd_numbers, RSpecOddNumberImplementation
+    to_list do
+      filter :force_error
+    end
+    public_actions :list
+  end
+end
+
+################################################################################
+#
+# Define our service, that implements both resources
+#
+class RSpecRemoteNumberService < Hoodoo::Services::Service
+  comprised_of RSpecOddNumberInterface
+end
+
 
 
 ##############################################################################
@@ -13,7 +227,7 @@ describe Hoodoo::Client do
 
     # Start our services in background threads
     spec_helper_start_svc_app_in_thread_for( RSpecNumberService )
-    spec_helper_start_svc_app_in_thread_for( RSpecRemoteNumberService)
+    spec_helper_start_svc_app_in_thread_for( RSpecRemoteNumberService )
 
   end
 
@@ -30,15 +244,15 @@ describe Hoodoo::Client do
 
     let( :resources ) { [
       {
-        endpoint:     @client.resource( :RSpecNumber, 1),
+        endpoint:     @client.resource( :RSpecNumber, 1 ),
         data:         (0..999).to_a
       },
       {
-        endpoint:     @client.resource( :RSpecEvenNumber, 1),
+        endpoint:     @client.resource( :RSpecEvenNumber, 1 ),
         data:         (0..999).step(2).to_a
       },
       {
-        endpoint:     @client.resource( :RSpecOddNumber, 1),
+        endpoint:     @client.resource( :RSpecOddNumber, 1 ),
         data:         (1..999).step(2).to_a
       },
     ] }
@@ -63,7 +277,7 @@ describe Hoodoo::Client do
 
       let(:limits) {
         # Note: Smaller limits will make the tests very slooooow
-        [ 25, 250, 500, 750, 999, 1000, 1001 ]
+        [ 1, 25, 250, 500, 750, 999, 1000, 1001 ]
       }
 
       it 'enumerates correctly with different batch sizes' do
@@ -92,23 +306,23 @@ describe Hoodoo::Client do
 
     let( :resources ) { [
       {
-        endpoint:         @client.resource( :RSpecNumber, 1),
+        endpoint:         @client.resource( :RSpecNumber, 1 ),
         expected_results: [
                             {
                               limit:  10,
-                              data:   (0..499).to_a,
+                              data:   ( 0..499 ).to_a,
                             },
                             {
                               limit:  250,
-                              data:   (0..499).to_a,
+                              data:   ( 0..499 ).to_a,
                             },
                             {
                               limit:  499,
-                              data:   (0..498).to_a,
+                              data:   ( 0..498 ).to_a,
                             },
                             {
                               limit:  500,
-                              data:   (0..499).to_a,
+                              data:   ( 0..499 ).to_a,
                             },
                             {
                               limit:  501,
@@ -117,19 +331,19 @@ describe Hoodoo::Client do
                           ]
       },
       {
-        endpoint:         @client.resource( :RSpecEvenNumber, 1),
+        endpoint:         @client.resource( :RSpecEvenNumber, 1 ),
         expected_results: [
                             {
                               limit:  10,
-                              data:   (0..498).step(2).to_a
+                              data:   ( 0..498 ).step(2).to_a
                             },
                             {
                               limit:  249,
-                              data:   (0..496).step(2).to_a
+                              data:   ( 0..496 ).step(2).to_a
                             },
                             {
                               limit:  250,
-                              data:   (0..498).step(2).to_a
+                              data:   ( 0..498 ).step(2).to_a
                             },
                             {
                               limit:  251,
@@ -138,19 +352,19 @@ describe Hoodoo::Client do
                           ]
       },
       {
-        endpoint:         @client.resource( :RSpecOddNumber, 1),
+        endpoint:         @client.resource( :RSpecOddNumber, 1 ),
         expected_results: [
                             {
                               limit:  10,
-                              data:   (1..499).step(2).to_a
+                              data:   ( 1..499 ).step(2).to_a
                             },
                             {
                               limit:  249,
-                              data:   (1..497).step(2).to_a
+                              data:   ( 1..497 ).step(2).to_a
                             },
                             {
                               limit:  250,
-                              data:   (1..499).step(2).to_a
+                              data:   ( 1..499 ).step(2).to_a
                             },
                             {
                               limit:  251,
@@ -202,9 +416,9 @@ describe Hoodoo::Client do
     it 'raises an exception if no block supplied' do
 
       endpoints = [
-          @client.resource( :RSpecNumber, 1),
-          @client.resource( :RSpecEvenNumber, 1),
-          @client.resource( :RSpecOddNumber, 1),
+          @client.resource( :RSpecNumber, 1 ),
+          @client.resource( :RSpecEvenNumber, 1 ),
+          @client.resource( :RSpecOddNumber, 1 ),
       ]
 
       endpoints.each do | endpoint |